% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaysSinceWhen.R
\name{DaysSinceWhen}
\alias{DaysSinceWhen}
\title{DaysSinceWhen}
\usage{
DaysSinceWhen(dob)
}
\arguments{
\item{dob}{A character string in the form "yyyy-mm-dd"}
}
\value{
a numeric decimal numeric value
}
\description{
Computes the number of days that have passed since
    the date entered.
}
\examples{
DaysOld <- DaysSinceWhen("1948-6-6")
}
\seealso{
\code{\link{Sys.time}}
}
